change integer to long

+++ src/main/java/org/gridsuite/study/server/StudyController.java
@@ -1688,7 +1688,7 @@ public class StudyController {
     @PostMapping(value = "/studies/{studyUuid}/sensitivity-analysis/factors-count")
     @Operation(summary = "Get the factors count of sensitivity parameters")
     @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The factors count of sensitivity parameters")})
-    public ResponseEntity<Integer> getSensitivityAnalysisFactorsCount(
+    public ResponseEntity<Long> getSensitivityAnalysisFactorsCount(
             @PathVariable("studyUuid") UUID studyUuid,
             @Parameter(description = "Is Injections Set") @RequestParam(name = "isInjectionsSet", required = false) Boolean isInjectionsSet,
             @RequestBody Map<String, List<UUID>> ids) {
